# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Astrade API",
        "description": "Astrade API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/astrade/public",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks with subtasks for the authenticated user",
                "description": "Get all tasks and their subtasks for the authenticated user.",
                "operationId": "7eb70d8d671c33c68394b46c6494a1b0",
                "responses": {
                    "200": {
                        "description": "List of tasks with subtasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/today": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get today's tasks for the authenticated user",
                "description": "Returns the tasks scheduled for today for the authenticated user.",
                "operationId": "63a783b546a4b4dd1c317b0d6486f1e9",
                "responses": {
                    "200": {
                        "description": "List of today's tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/planned/{days}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get planned tasks for the next N days",
                "description": "Returns the tasks planned for the next N days (maximum 30).",
                "operationId": "63456f554584da75cd18e9c820a1b40f",
                "parameters": [
                    {
                        "name": "days",
                        "in": "path",
                        "description": "Number of days to look ahead (max 30)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of planned tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/{task_id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get details of a specific task",
                "description": "Displays the details of a task for the authenticated user.",
                "operationId": "bd69e9b15de99bcda89565ecb76a1337",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task details with subtasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/status/summary": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a summary of task statuses",
                "description": "Returns a summary of the status of the user's tasks.",
                "operationId": "85ec4e1c4649a074a948b3a16c004018",
                "responses": {
                    "200": {
                        "description": "Summary of completed and pending tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/{taskId}/companies": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get company associated with a specific task",
                "description": "View companies associated with a specific task via URL param.",
                "operationId": "5c8b7005b9a9754f8c48f87586670f59",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found or not authorized"
                    },
                    "400": {
                        "description": "No company associated with this task"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/by-date/{date}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks for a specific date",
                "description": "Returns the tasks scheduled for a specific date for the authenticated user.",
                "operationId": "af33e2995c29fa37d93f7958a4e6d94b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Date in YYYY-MM-DD format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-09-25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks for the given date",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid date format"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/day-offset/{offset}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get tasks for a specific day offset",
                "description": "Returns the tasks for a specific day offset for the authenticated user.",
                "operationId": "3b73a426b1658c80ca7d70860a1dfdd9",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "path",
                        "description": "Number of days from today (0 = today)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks for the given day",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid offset"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "success": {
                        "description": "Returns a standardized JSON response",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}